requirements.txt
# =============================================================================
# GLM-OCR Production Service — Dependencies
# =============================================================================

# --- Core ---
pyyaml>=6.0
pydantic>=2.0
pydantic-settings>=2.0

# --- PDF Processing ---
PyMuPDF>=1.24.0              # fitz — fastest PDF rasterizer
Pillow>=10.0

# --- DOC/DOCX Processing ---
python-docx>=1.1.0
mammoth>=1.8.0               # .doc → HTML/Markdown

# --- OCR Model Client ---
openai>=1.30.0               # OpenAI-compatible client for vLLM/SGLang
httpx>=0.27.0                # async HTTP
aiohttp>=3.9.0

# --- Layout Detection ---
# PP-DocLayout-V3 (optional — install separately)
# paddlepaddle-gpu>=2.6.0
# paddlex>=2.0.0
onnxruntime-gpu>=1.18.0      # for layout model ONNX inference
numpy>=1.26.0

# --- API Server ---
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9      # file upload support

# --- Utilities ---
tqdm>=4.66.0
aiofiles>=24.1.0

# --- GLM-OCR Official SDK (optional, can use directly) ---
# glmocr

# --- Inference Backends (install on the server separately) ---
# vllm (nightly): pip install -U vllm --extra-index-url https://wheels.vllm.ai/nightly
# sglang: pip install sglang[all]
# transformers (from source): pip install git+https://github.com/huggingface/transformers.git
